Moving tests to their folder and decopling from django
All test files were moved to the "tests" directory, the tasks file was
removed, since it was only bothering making us to add import statements
in crazy places.

Also, a *temporary* conf file was added until we come up with a better
solution that allows us easily plugging this library in django apps.